+++ src/canvas/Canvas.ts
@@ -1395,6 +1395,9 @@ export class Canvas extends SelectableCanvas {
 
     if (actionHandler) {
       actionPerformed = actionHandler(e, transform, x, y);
+      // update lastX/Y
+      transform.lastX = x;
+      transform.lastY = y;
     }
     if (action === 'drag' && actionPerformed) {
       transform.target.isMoving = true;