+++ src/controls/wrapWithFixedAnchor.ts
@@ -1,20 +1,50 @@
-import { Transform, TransformActionHandler } from '../EventTypeDefs';
+import {
+  Transform,
+  TransformAction,
+  TransformActionHandler,
+} from '../EventTypeDefs';
+import { TOriginX, TOriginY } from '../typedefs';
 
 /**
  * Wrap an action handler with saving/restoring object position on the transform.
  * this is the code that permits to objects to keep their position while transforming.
- * @param {Function} actionHandler the function to wrap
- * @return {Function} a function with an action handler signature
+ * @param actionHandler the function to wrap
+ * @param [originAdaptor] enables adapting origin of the anchor point
+ * @return wrapped function
  */
 export function wrapWithFixedAnchor<T extends Transform>(
-  actionHandler: TransformActionHandler<T>
-) {
-  return ((eventData, transform, x, y) => {
-    const { target, originX, originY } = transform,
-      centerPoint = target.getRelativeCenterPoint(),
+  actionHandler: TransformActionHandler<T>,
+  originAdaptor?: TransformAction<
+    T,
+    {
+      x: TOriginX;
+      y: TOriginY;
+    }
+  >,
+  originAdaptorAfter?: TransformAction<
+    T,
+    {
+      x: TOriginX;
+      y: TOriginY;
+    }
+  >
+): TransformActionHandler<T> {
+  return (eventData, transform, x, y) => {
+    const {
+        target,
+        originX: incomingOriginX,
+        originY: incomingOriginY,
+      } = transform,
+      { x: originX, y: originY } = originAdaptor
+        ? originAdaptor(eventData, transform, x, y)
+        : { x: incomingOriginX, y: incomingOriginY };
+    const centerPoint = target.getRelativeCenterPoint(),
       constraint = target.translateToOriginPoint(centerPoint, originX, originY),
       actionPerformed = actionHandler(eventData, transform, x, y);
-    target.setPositionByOrigin(constraint, originX, originY);
+    const { x: originXAfter, y: originYAfter } = originAdaptorAfter
+      ? originAdaptorAfter(eventData, transform, x, y)
+      : { x: originX, y: originY };
+    target.setPositionByOrigin(constraint, originXAfter, originYAfter);
     return actionPerformed;
-  }) as TransformActionHandler<T>;
+  };
 }