+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - fix(Geometry): `_getCoords` not respecting group [#8747](https://github.com/fabricjs/fabric.js/issues/8747)
+- feat(): Controls adaptors [#8514](https://github.com/fabricjs/fabric.js/pull/8514)
 - ci(): disallow circular deps [#8759](https://github.com/fabricjs/fabric.js/pull/8759)
 - fix(): env WebGL import cycle [#8758](https://github.com/fabricjs/fabric.js/pull/8758)
 - chore(TS): remove controls from prototype. BREAKING: controls aren't shared anymore [#8753](https://github.com/fabricjs/fabric.js/pull/8753)